requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
DateTime==4.3
greenlet==1.1.2
idna==3.3
numpy==1.21.5
pandas==1.3.5
psycopg2==2.9.2
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
urllib3==1.26.7
zope.interface==5.4.0